pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axis-pia-project"
version = "0.1.0"
description = "LangGraph-based router with calculator and email agents"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.1.0",
    "litellm>=1.0.0",
    "langchain-litellm>=0.1.0",
    "langchain-community>=0.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]
